Add support for multi-character wildcard '*'
I was using FastParse for both parsing the source into a Glob, and for
testing input against the Glob.

As it turns out, FastParse isn't well equipped for backtracking across
parsers, so the glob for a dynamically generated parser for wildcard
then suffix (ex: "*string") wasn't within it's feature set.

This commit also converts to using Regex for testing input against a Glob.